% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{arr}
\alias{arr}
\title{Abbreviated scalar}
\usage{
arr(data = NA, dim = length(data), dimnames = NULL)
}
\arguments{
\item{data}{a vector (including a list or \code{\link[base]{expression}}
    vector) giving data to fill the array.  Non-atomic classed objects
    are coerced by \code{\link[base]{as.vector}}.}

\item{dim}{the dim attribute for the array to be created, that is an
    integer vector of length one or more giving the maximal indices in
    each dimension.}

\item{dimnames}{either \code{NULL} or the names for the dimensions.
    This must a list (or it will be ignored) with one component for each
    dimension, either \code{NULL} or a character vector of the length
    given by \code{dim} for that dimension.  The list can be named, and
    the list names will be used as names for the dimensions.  If the
    list is shorter than the number of dimensions, it is extended by
    \code{NULL}s to the length required.}
}
\description{
Abbreviated scalar
}
