% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{chrsplit}
\alias{chrsplit}
\title{strsplit}
\usage{
chrsplit(
  x,
  split = "",
  unlist = TRUE,
  fixed = FALSE,
  perl = FALSE,
  use_bytes = FALSE
)
}
\arguments{
\item{x}{
    character vector, each element of which is to be split.  Other
    inputs, including a factor, will give an error.
  }

\item{split}{
    character vector (or object which can be coerced to such)
    containing \link[base]{regular expression}(s) (unless \code{fixed = TRUE})
    to use for splitting.  If empty matches occur, in particular if
    \code{split} has length 0, \code{x} is split into single characters.
    If \code{split} has length greater than 1, it is re-cycled along
    \code{x}.
  }

\item{unlist}{logical. Should the result ve a vector?}

\item{fixed}{
    logical.  If \code{TRUE} match \code{split} exactly, otherwise
    use regular expressions.  Has priority over \code{perl}.
  }

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{use_bytes}{logical. If TRUE the matching is done byte-by-byte rather than character-by-character, and inputs with marked encodings are not converted.}
}
\description{
strsplit
}
