% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{cnms}
\alias{cnms}
\title{colnames, but consistent with terminology}
\usage{
cnms(x, do_NULL = TRUE, prefix = "col")
}
\arguments{
\item{x}{a matrix-like \R object, with at least two dimensions for
    \code{colnames}.}

\item{do_NULL}{logical. If FALSE and names are NULL, names are created.}

\item{prefix}{for created names.}
}
\description{
colnames, but consistent with terminology
}
