% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{:=}
\alias{:=}
\title{Ergo protoype kickstarter - wrapper for \code{ergo::tv()}.}
\usage{
`:=`(x)
}
\arguments{
\item{x}{an expression.}
}
\description{
Ergo protoype kickstarter - wrapper for \code{ergo::tv()}.
}
