% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{control_flow}
\alias{control_flow}
\title{Control flow component for ? operator}
\usage{
control_flow(query, ...)
}
\arguments{
\item{query}{query}

\item{...}{conditions}
}
\description{
Control flow component for ? operator
}
