% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_order.R
\name{do_fil}
\alias{do_fil}
\title{Filter, but humanly readable}
\usage{
do_fil(x, f)
}
\arguments{
\item{x}{a vector.}

\item{f}{a function of the appropriate arity (binary for
    \code{Reduce}, unary for \code{Filter}, \code{Find} and
    \code{Position}, \eqn{k}-ary for \code{Map} if this is called with
    \eqn{k} arguments).  An arbitrary predicate function for
    \code{Negate}.}
}
\description{
Filter, but humanly readable
}
