% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_order.R
\name{do_pos}
\alias{do_pos}
\title{Position, but humanly readable}
\usage{
do_pos(f, x, right = FALSE, no_match = NA_integer_)
}
\arguments{
\item{f}{a function of the appropriate arity (binary for
    \code{Reduce}, unary for \code{Filter}, \code{Find} and
    \code{Position}, \eqn{k}-ary for \code{Map} if this is called with
    \eqn{k} arguments).  An arbitrary predicate function for
    \code{Negate}.}

\item{x}{a vector.}

\item{right}{a logical indicating whether to proceed from left to
    right (default) or from right to left.}

\item{no_match}{the value to be returned in the case when “no match”
(no element satisfying the predicate) is found.}
}
\description{
Position, but humanly readable
}
