% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_order.R
\name{do_red}
\alias{do_red}
\title{Reduce but humanly readable}
\usage{
do_red(x, f, init, right = FALSE, accumulate = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{f}{a function of the appropriate arity (binary for
    \code{Reduce}, unary for \code{Filter}, \code{Find} and
    \code{Position}, \eqn{k}-ary for \code{Map} if this is called with
    \eqn{k} arguments).  An arbitrary predicate function for
    \code{Negate}.}

\item{init}{an \R object of the same kind as the elements of
    \code{x}.}

\item{right}{a logical indicating whether to proceed from left to
    right (default) or from right to left.}

\item{accumulate}{a logical indicating whether the successive reduce
    combinations should be accumulated.  By default, only the final
    combination is used.}
}
\description{
Reduce but humanly readable
}
