% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dup}
\alias{dup}
\title{Abbreviated duplicated.}
\usage{
dup(x, incomparables = FALSE)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}

\item{incomparables}{a vector of values that cannot be compared.
    \code{FALSE} is a special value, meaning that all values can be
    compared, and may be the only value accepted for methods other than
    the default.  It will be coerced internally to the same type as
       \code{x}.}
}
\description{
Abbreviated duplicated.
}
