% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{full}
\alias{full}
\title{Make abbreviations into full nouns.}
\usage{
full(x, ...)
}
\arguments{
\item{x}{string from formula in \verb{?}}

\item{...}{dots}
}
\description{
Make abbreviations into full nouns.
}
\keyword{internal}
