% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{get_function}
\alias{get_function}
\title{Make abbreviations into full nouns.}
\usage{
get_function(fun, prepend)
}
\arguments{
\item{x}{string from formula in \verb{?}}

\item{...}{dots}
}
\description{
Make abbreviations into full nouns.
}
\keyword{internal}
