% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{?_controlflow}
\alias{?_controlflow}
\title{Control flow component for ? operator}
\usage{
`?_controlflow`(...)
}
\arguments{
\item{...}{conditions}
}
\description{
Control flow component for ? operator
}
