% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{int}
\alias{int}
\title{Abbreviated scalar}
\usage{
int(length = 0)
}
\arguments{
\item{length}{A non-negative integer specifying the desired length.
    Double values will be coerced to integer:
    supplying an argument of length other than one is an error.}
}
\description{
Abbreviated scalar
}
