% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{lap}
\alias{lap}
\title{lapply, but ergonomic}
\usage{
lap(x, fun, ...)
}
\arguments{
\item{x}{vector or an expression or coercible by \code{as.list}.}

\item{fun}{a function to be applied to each element.}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
lapply, but ergonomic
}
\keyword{list}
