% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_tv}
\alias{load_tv}
\title{abbreviations of common tidyverse nouns for quick prototyping}
\usage{
load_tv()
}
\description{
abbreviations of common tidyverse nouns for quick prototyping
}
\keyword{internal}
