% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{for-loop prototyping}
\usage{
loop(expr, ..., type = "along")
}
\description{
for-loop prototyping
}
\examples{
\dontrun{
x = as.list(1:5)
loop({
 y[[i]] <- x[[i]] + 1
}, x)

x = mtx(1:9, ncol = 3)
loop({
 y[i] <- sum(x[i,])
}, x, type = "row")
}
}
