% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{lst}
\alias{lst}
\title{Abbreviated scalar}
\usage{
lst(..., names = NULL)
}
\arguments{
\item{...}{objects, possibly named.}

\item{names}{a character vector of list element names}
}
\description{
Abbreviated scalar
}
