% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{mlgl}
\alias{mlgl}
\title{grepl, but wrapped for consistency}
\usage{
mlgl(
  x,
  pattern,
  ignore_case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  use_bytes = FALSE
)
}
\arguments{
\item{x}{a character vector where matches are sought, or an
    object which can be coerced by \code{as.character} to a character
    vector.  \link[base]{Long vectors} are supported.}

\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr}, \code{gregexpr} and \code{regexec}.}

\item{ignore_case}{should lettercase be ignored?}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{fixed}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}

\item{use_bytes}{logical. If TRUE the matching is done byte-by-byte rather than character-by-character, and inputs with marked encodings are not converted.}
}
\description{
grepl, but wrapped for consistency
}
