% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{mpos}
\alias{mpos}
\title{grep, but wrapped for consistency}
\usage{
mpos(
  x,
  pattern,
  ignore_case = FALSE,
  perl = FALSE,
  value = FALSE,
  fixed = FALSE,
  use_bytes = FALSE,
  invert = FALSE
)
}
\arguments{
\item{x}{a character vector where matches are sought, or an
    object which can be coerced by \code{as.character} to a character
    vector.  \link[base]{Long vectors} are supported.}

\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr}, \code{gregexpr} and \code{regexec}.}

\item{ignore_case}{should lettercase be ignored?}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{value}{if \code{FALSE}, a vector containing the (\code{integer})
    indices of the matches determined by \code{grep} is returned, and if
    \code{TRUE}, a vector containing the matching elements themselves is
    returned.}

\item{fixed}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}

\item{use_bytes}{should bytes be used?}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}
}
\description{
grep, but wrapped for consistency
}
