% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{msub}
\alias{msub}
\title{sub but named wrapped for consistency}
\usage{
msub(
  x,
  pattern,
  replacement,
  ignore_case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  use_bytes = FALSE
)
}
\arguments{
\item{x}{a character vector where matches are sought, or an
    object which can be coerced by \code{as.character} to a character
    vector.  \link[base]{Long vectors} are supported.}

\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr}, \code{gregexpr} and \code{regexec}.}

\item{replacement}{a replacement for matched pattern in \code{sub} and
    \code{gsub}.  Coerced to character if possible.  For
    \code{fixed = FALSE} this can include backreferences \code{"\1"} to
    \code{"\9"} to parenthesized subexpressions of \code{pattern}.  For
    \code{perl = TRUE} only, it can also contain \code{"\U"} or
    \code{"\L"} to convert the rest of the replacement to upper or
    lower case and \code{"\E"} to end case conversion.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  If \code{NA}, all elements in the result
    corresponding to matches will be set to \code{NA}.
  }

\item{ignore_case}{should lettercase be ignored?}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{fixed}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}

\item{use_bytes}{should bytes be used?}
}
\description{
sub but named wrapped for consistency
}
