% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{mtx}
\alias{mtx}
\title{Abbreviated scalar}
\usage{
mtx(data = NA, nrow = 1, ncol = 1, by_row = FALSE, dimnames = NULL)
}
\arguments{
\item{data}{an optional data vector (including a list or
    \code{\link[base]{expression}} vector).  Non-atomic classed \R objects are
    coerced by \code{\link[base]{as.vector}} and all attributes discarded.}

\item{nrow}{the desired number of rows.}

\item{ncol}{the desired number of columns.}

\item{by_row}{logical. If FALSE (the default) the matrix is filled by columns, otherwise the matrix is filled by rows.}

\item{dimnames}{A \code{\link[base]{dimnames}} attribute for the matrix:
    \code{NULL} or a \code{list} of length 2 giving the row and column
    names respectively.  An empty list is treated as \code{NULL}, and a
    list of length one as row names.  The list can be named, and the
    list names will be used as names for the dimensions.}
}
\description{
Abbreviated scalar
}
