% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{p0}
\alias{p0}
\title{paste0 but concise}
\usage{
p0(..., collapse = "", recycle0 = FALSE)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{recycle0}{\code{\link[base]{logical}} indicating if zero-length
    character arguments should lead to the zero-length
    \code{\link[base]{character}(0)} after the \code{sep}-phase (which turns into
    \code{""} in the \code{collapse}-phase, i.e., when \code{collapse} is
    not \code{NULL}).}
}
\description{
paste0 but concise
}
