% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{rnms}
\alias{rnms}
\title{rownames, but consistent with terminology}
\usage{
rnms(x, do_NULL = TRUE, prefix = "row")
}
\arguments{
\item{x}{a matrix-like \R object, with at least two dimensions for
    \code{colnames}.}

\item{do_NULL}{logical. If FALSE and names are NULL, names are created.}

\item{prefix}{for created names.}
}
\description{
rownames, but consistent with terminology
}
