% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{rsum}
\alias{rsum}
\title{rowSums, but consistent with terminology}
\usage{
rsum(x, na_rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric,
    complex, integer or logical values, or a numeric data frame.  For
    \code{.colSums()} etc, a numeric, integer or logical matrix (or
    vector of length \code{m * n}).}

\item{na_rm}{Boolean. Remove missing values?}

\item{dims}{integer: Which dimensions are regarded as \sQuote{rows} or
    \sQuote{columns} to sum over.  For \code{row*}, the sum or mean is
    over dimensions \code{dims+1, \dots}; for \code{col*} it is over
    dimensions \code{1:dims}.}
}
\description{
rowSums, but consistent with terminology
}
