% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{sap}
\alias{sap}
\title{sapply, but ergonomic}
\usage{
sap(x, fun, ..., simplify = TRUE, use_names = TRUE)
}
\arguments{
\item{x}{vector or an expression or coercible by \code{as.list}.}

\item{fun}{a function to be applied to each element.}

\item{...}{dynamic dots}

\item{simplify}{logical or character string; should the result be
    simplified to a vector, matrix or higher dimensional array if
    possible?  For \code{sapply} it must be named and not abbreviated.
    The default value, \code{TRUE}, returns a vector or matrix if appropriate,
    whereas if \code{simplify = "array"} the result may be an
    \code{\link[base]{array}} of \dQuote{rank}
    (\eqn{=}\code{length(dim(.))}) one higher than the result
    of \code{FUN(X[[i]])}.}

\item{use_names}{should the output contain names?}
}
\description{
sapply, but ergonomic
}
