% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seed}
\alias{seed}
\title{Setting seed}
\usage{
seed(seed, kind = NULL, normal_kind = NULL, sample_kind = NULL)
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}

\item{kind}{character or \code{NULL}.  If \code{kind} is a character
    string, set \R's RNG to the kind desired.  Use \code{"default"} to
    return to the \R default.  See \sQuote{Details} for the
    interpretation of \code{NULL}.}

\item{normal_kind}{character string or NULL. If it is a character string,
set the method of Normal generation. Use "default" to return to the R
default. NULL makes no change.}

\item{sample_kind}{character string or NULL. If it is a character string,
set the method of discrete uniform generation (used in sample, for
instance). Use "default" to return to the R default. NULL makes no change.}
}
\description{
Setting seed
}
