% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{tv}
\alias{tv}
\title{rapid tidyverse prototyping}
\usage{
tv(expr)
}
\arguments{
\item{expr}{an expression.}
}
\description{
rapid tidyverse prototyping
}
\examples{
\dontrun{
iris \%>\%
  m(n = 1:n()) \%>\%
  f(Sepal.Length > 5) |>
  tv()
}
}
