% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{vap}
\alias{vap}
\title{vapply, but ergonomic}
\usage{
vap(x, fun, fun_value, ..., use_names = TRUE)
}
\arguments{
\item{x}{vector or an expression or coercible by \code{as.list}.}

\item{fun}{a function to be applied to each element.}

\item{fun_value}{a template for the return value from fun.}

\item{...}{optional arguments to \code{FUN}.}

\item{use_names}{should the output contain names?}
}
\description{
vapply, but ergonomic
}
