% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vec}
\alias{vec}
\title{Abbreviated scalar}
\usage{
vec(mode = "logical", length = 0)
}
\arguments{
\item{mode}{character string naming an atomic mode or
    \code{"list"} or \code{"expression"} or (except for \code{vector})
    \code{"any"}.  Currently, \code{is.vector()} allows any type (see
    \code{\link[base]{typeof}}) for \code{mode}, and when mode is not
    \code{"any"}, \code{is.vector(x, mode)} is almost the same as
    \code{typeof(x) == mode}.}

\item{length}{a non-negative integer specifying the desired length.  For
    a \link[base]{long vector}, i.e., \code{length > .Machine$integer.max}, it
    has to be of type \code{"double"}.  Supplying an argument of length
    other than one is an error.}
}
\description{
Abbreviated scalar
}
