% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{?_controlflow}
\alias{?_controlflow}
\title{Control flow component for ? operator}
\usage{
`?_controlflow`(x, y)
}
\arguments{
\item{x}{a logical vector or scalar.}

\item{y}{a vector, scalar or compound expression using :.}
}
\description{
Control flow component for ? operator
}
