% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{?}
\alias{?}
\title{binary ? can now used for typechecking in control flow}
\usage{
`?`(x, y)
}
\description{
binary ? can now used for typechecking in control flow
}
\examples{
\dontrun{
x = as.character(1:5)
# is x a character vector?
x ? chr
# if x is character, make it an integer
# old: 'if(is.character(x)) as.integer(x) else x
if(x?chr) (x?~int) else x
}
}
