% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{?}
\alias{?}
\title{binary ? can now used for type checking in control flow}
\usage{
`?`(x, y)
}
\arguments{
\item{x}{scalar, vector, name or call.}

\item{y}{scalar or vector.}
}
\value{
If conditional input, a vector or scalar. If help is sought, silent.
}
\description{
The package?packagename functionality is dropped in favor of concise
type checking. The ? occupies valuable real estate on the keyboard, and is
given high operator precedence in R. It can then be used to make chain actions
that make code much more concise yet humanly readable.
}
\examples{
\dontrun{
x = as.character(1:5)
# is x a character vector?
x ? chr
# if x is character, make it an integer
# old: 'if(is.character(x)) as.integer(x) else x
if(x?chr) (x?~int) else x
}
}
