% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{for-loop prototyping}
\usage{
loop(expr, ..., type = "along")
}
\arguments{
\item{expr}{an expression of the body of a loop.}

\item{...}{dynamic dots that include objects necessary for the body of the loop.}

\item{type}{the type of loop to construct. along for a sequence alonga vector, row for a sequence along the rows of an object.}
}
\description{
for-loop prototyping
}
\examples{
x = as.list(1:5)
loop({ y[[i]] <- x[[i]] + 1}, x)
x = mtx(1:9, ncol = 3)
loop({y[i] <- sum(x[i,])}, x, type = "row")
}
